+++ example/lib/common.dart
@@ -1,0 +1,25 @@
+import 'dart:developer';
+
+import 'package:http_interceptor/http_interceptor.dart';
+
+class LoggerInterceptor extends InterceptorContract {
+  @override
+  Future<BaseRequest> interceptRequest({
+    required BaseRequest request,
+  }) async {
+    log('----- Request -----');
+    log(request.toString());
+    log(request.headers.toString());
+    return request;
+  }
+
+  @override
+  Future<BaseResponse> interceptResponse({
+    required BaseResponse response,
+  }) async {
+    log('----- Response -----');
+    log('Code: ${response.statusCode}');
+    log(response.toString());
+    return response;
+  }
+}