+++ lib/http_interceptor.dart
@@ -1,6 +1,21 @@
 library http_interceptor;
 
 export 'package:http/http.dart';
-export 'package:http_interceptor/extensions/extensions.dart';
-export 'package:http_interceptor/http/http.dart';
-export 'package:http_interceptor/models/models.dart';
+
+export './extensions/base_request.dart';
+export './extensions/base_response_none.dart'
+    if (dart.library.io) './extensions/base_response_io.dart';
+export './extensions/multipart_request.dart';
+export './extensions/request.dart';
+export './extensions/request.dart';
+export './extensions/response.dart';
+export './extensions/streamed_request.dart';
+export './extensions/streamed_response.dart';
+export './extensions/string.dart';
+export './extensions/uri.dart';
+export './http/http_methods.dart';
+export './http/intercepted_client.dart';
+export './http/intercepted_http.dart';
+export './models/http_interceptor_exception.dart';
+export './models/interceptor_contract.dart';
+export './models/retry_policy.dart';