+++ lib/models/http_interceptor_exception.dart
@@ -1,8 +1,9 @@
 class HttpInterceptorException implements Exception {
-  final message;
+  final dynamic message;
 
   HttpInterceptorException([this.message]);
 
+  @override
   String toString() {
     if (message == null) return "Exception";
     return "Exception: $message";